requirements.txt
Flask==1.0.3
flask-cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
faker==1.0.7
google-cloud-storage==1.16.0
python-dotenv==0.10.2
flask_jwt_extended==3.18.2
flask-migrate==2.5.2